Change creation of encryption VFSs
Up to now a single VFS for encryption was created based on the (current) default VFS for a platform. This approach made it difficult to use non-default VFSs in combination with database encryption. From now on it is possible to create more than one VFS for encryption combining it with different underlying VFSs. These combinations can be used in parallel.

Additionally, the version information has been placed in a separate C header file.