Modify search for multiple ciphers VFS
Function mcFindVfs uses now a special op code, SQLITE3MC_FCNTL_PVFS, in a call to sqlite3_file_control to determine the address of the multiple ciphers VFS. As soon as a multiple ciphers VFS has been found on the VFS stack, the new op code will not be forwarded further. Only the top-most multiple ciphers VFS on the stack will actually do any encryption. That is, adding the multiple ciphers VFS to the VFS stack multiple times is harmless.