Disable cipher scheme AEGIS for MSVC 2015 and below (issue #195)
MSVC 2015 is very strict regarding data alignment. Passing aligned parameters by value is not supported. Therefore the cipher scheme AEGIS is disabled for now.

The underlying libaegis library passes aligned parameters by value for a number of functions. To support older MSVC compilers it would be necessary to change the parameter passing mechanism for the affected functions from "by value" to "by reference". In case this will done in the upstream libaegis library, the AEGIS cipher scheme will be re-enabled for MSVC 2015 and below.