Fix issue #221
In function sqlite3mc_initialize() the encryption VFS shim was added to the default VFS unconditionally. However, this fails if no default VFS is configured. Now it is first checked whether a default VFS is configured. Only then the encryption VFS shim is applied, so that sqlite3mc_initialize() does not fail.